% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PartialDistance}
\alias{PartialDistance}
\title{Compute the partial distance between matrices}
\usage{
PartialDistance(Ar, Br)
}
\arguments{
\item{Ar}{A numeric n-by-m matrix containing the position of n data points m-dimensional points}

\item{Br}{A numeric k-by-m matrix containing the position of k reference m-dimensional points}
}
\value{
A numeric n-by-k matrix reporting the euclidean distance between the n data points and 
the k reference points
}
\description{
This function computes the distance between a matrix of data points and a
matrix of reference points. Both matrices need to have the same number of
dimensions (columns). The code was taked from here: // C++ code taken from here:
https://www.r-bloggers.com/pairwise-distances-in-r/
}
\examples{

A <- matrix(runif(10000*100), nrow = 10000)
B <- matrix(runif(100*100), nrow = 100)

library(distutils)

print(system.time(C1 <- PartialDistance(A, B)))
print(system.time(C2 <- as.matrix(dist(rbind(A,B)))[1:10000, 10001:10100]))

summary(as.vector(C1 - C2))


}
