% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Partition}
\alias{Partition}
\title{Parition a set of data ponints into groups based on the distance from a set of reference points}
\usage{
Partition(Ar, Br, SquaredAr)
}
\arguments{
\item{Ar}{A numeric n-by-m matrix containing the position of n data points m-dimensional points}

\item{Br}{A numeric k-by-m matrix containing the position of k reference m-dimensional points}

\item{SquaredAr}{A numeric vector reporting the sum, by row, of Ar (i.e. SquaredX = rowSums(Ar^2)).
This is done to speed up the calculation when the same set o data points is clustered against
a different set of reference point (e.g. in k-means)}
}
\value{
@details list with two elements
\itemize{
\item Patition is an integer vector indicating, for each point, the index of the closest reference point
\item Distance is a numeric vector indicating, for each point, the squared distance to the closest reference point
}
}
\description{
This function computes the distance between a matrix of data points and a
matrix of reference points. Then associate each point with a id inicating the closet
reference point
}
\examples{

A <- matrix(runif(10000*100), nrow = 10000)
B <- matrix(runif(100*100), nrow = 100)
  
  library(distutils)
  
  print(
    system.time(
      Part <- Partition(Ar = A, Br = B, SquaredAr = rowSums(A^2))
    )
  )
  
  print(
    system.time({
      C2 <- as.matrix(dist(rbind(A,B)))[1:10000, 10001:10100]
      Partition <- apply(C2, 1, which.min)
      Dist <- apply(C2, 1, min)^2
    })
  )
  
  summary(Partition - Part$Patition)
  summary(Dist - Part$Dist)

}
