% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PenalizedElasticEnergy}
\alias{PenalizedElasticEnergy}
\title{Compute the penalized elastic energy associated with a particular configuration}
\usage{
PenalizedElasticEnergy(X, NodePositions, ElasticMatrix, Dists, alpha, beta)
}
\arguments{
\item{X}{A numeric n-by-m matrix containing the position of n data points m-dimensional points}

\item{NodePositions}{A numeric k-by-m matrix containing the position of the k nodes of the embedded graph}

\item{ElasticMatrix}{A numeric l-by-l matrix containing the elastic parameters associates with the edge
of the embedded graph}

\item{Dists}{A numeric vector containind the squared distance of the data points to the closest node of the graph}

\item{alpha}{}

\item{beta}{}
}
\value{
A list with four elements:
\itemize{
\item ElasticEnergy is the total energy
\item EP is the EP component of the energy
\item RS is the RS component of the energy
\item MSE is the MSE component of the energy
}
}
\description{
This function computes the elastic energy associate to a set of points and graph embedded
into them. See XXX for reference
}
\examples{

}
