% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RadialCount}
\alias{RadialCount}
\title{Compute the number of points within a given distance}
\usage{
RadialCount(Ar, Pr, SquaredAr, DVect)
}
\arguments{
\item{Ar}{A numeric n-by-m matrix containing the position of n data points m-dimensional points}

\item{DVect}{A k dimensional vector matrix containing the k distances to consider}

\item{P}{A m dimensional vector matrix containing the position of the reference point}
}
\value{
A k-dimensional vector containing the number of points with a distance lower than k
}
\description{
This function computes the distance between a matrix of data points and a
single reference points. A sorted distance vector define the reference distances
to consider
}
\examples{

A <- matrix(runif(10000*100), nrow = 10000)
B <- matrix(runif(100*100), nrow = 100)

library(distutils)

print(system.time(C1 <- PartialDistance(A, B)))
print(system.time(C2 <- as.matrix(dist(rbind(A,B)))[1:10000, 10001:10100]))

summary(as.vector(C1 - C2))


}
